[Java] P ints should be extracted as Java Longs
Other backends use the widest integral types available on the system.

+++ Src/PRuntimes/PJavaRuntime/src/test/java/ValueCloneTest.java
@@ -4,7 +4,7 @@ import prt.exceptions.UncloneableValueException;
 
 
 import java.util.*;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static prt.values.Clone.deepClone;
@@ -23,7 +23,7 @@ public class ValueCloneTest {
         Boolean b = Boolean.valueOf(true);
         assertEquals(deepClone(b), b);
 
-        Integer i = Integer.valueOf(31337);
+        Long i = Long.valueOf(31337);
         assertEquals(deepClone(i), i);
 
         Float f = Float.valueOf(1.61803f);
@@ -37,20 +37,20 @@ public class ValueCloneTest {
     @DisplayName("Can clone lists")
     public void testCloneList() {
         // Ensure the clone completes successfully
-        ArrayList<Integer> a1 = new ArrayList<>(List.of(1,2,3,4,5));
-        ArrayList<Integer> a2 = (ArrayList<Integer>) deepClone(a1);
+        ArrayList<Long> a1 = new ArrayList<>(List.of(1L,2L,3L,4L,5L));
+        ArrayList<Long> a2 = (ArrayList<Long>) deepClone(a1);
         assertEquals(a1, a2);
 
         // Now reassign an element and ensure only structural equality
-        a1.set(1, 42);
+        a1.set(1, 42L);
         assertNotEquals(a1, a2);
-        a2.set(1, 42);
+        a2.set(1, 42L);
         assertEquals(a1, a2);
 
         // Now append an element and ensure only structural equality.
-        a1.add(99);
+        a1.add(99L);
         assertNotEquals(a1, a2);
-        a2.add(99);
+        a2.add(99L);
         assertEquals(a1, a2);
     }
 
@@ -58,32 +58,32 @@ public class ValueCloneTest {
     @DisplayName("Can clone sets")
     public void testCloneSet() {
         // Ensure the clone completes successfully
-        LinkedHashSet<Integer> s1 = new LinkedHashSet<>(List.of(1,2,3,4,5));
-        LinkedHashSet<Integer> s2 = (LinkedHashSet<Integer>) deepClone(s1);
+        LinkedHashSet<Long> s1 = new LinkedHashSet<>(List.of(1L,2L,3L,4L,5L));
+        LinkedHashSet<Long> s2 = (LinkedHashSet<Long>) deepClone(s1);
         assertEquals(s1, s2);
 
         // Now mutate an element and ensure only structural equality
-        s1.add(6);
+        s1.add(6L);
         assertNotEquals(s1, s2);
-        s2.add(6);
+        s2.add(6L);
         assertEquals(s1, s2);
     }
 
     @Test
     @DisplayName("Can clone maps")
     public void testCloneMap() {
-        HashMap<String, Integer> m1 = new HashMap<>(Map.of(
-                "A", 1,
-                "B", 2,
-                "C", 3));
-        HashMap<String, Integer> m2 = (HashMap<String, Integer>) deepClone(m1);
+        HashMap<String, Long> m1 = new HashMap<>(Map.of(
+                "A", 1L,
+                "B", 2L,
+                "C", 3L));
+        HashMap<String, Long> m2 = (HashMap<String, Long>) deepClone(m1);
         assertEquals(m1, m2);
 
 
         // Ensure structural equality under adding elements
-        m1.put("Z", 42);
+        m1.put("Z", 42L);
         assertNotEquals(m1, m2);
-        m2.put("Z", 42);
+        m2.put("Z", 42L);
         assertEquals(m1, m2);
 
         // Ensure structural equality under removing elements
@@ -93,41 +93,41 @@ public class ValueCloneTest {
         assertEquals(m1, m2);
 
         // Ensure structural equality under reassigning values
-        m1.put("C", 99);
+        m1.put("C", 99L);
         assertNotEquals(m1, m2);
-        m2.put("C", 99);
+        m2.put("C", 99L);
         assertEquals(m1, m2);
     }
 
     @Test
     @DisplayName("Can clone nested collections")
     public void testNestedCollections() {
-        HashMap<String, ArrayList<Integer>> m1 = new HashMap<>(Map.of(
-                "123", new ArrayList<>(List.of(1,2,3)),
-                "987", new ArrayList<>(List.of(9,8,7))
+        HashMap<String, ArrayList<Long>> m1 = new HashMap<>(Map.of(
+                "123", new ArrayList<>(List.of(1L,2L,3L)),
+                "987", new ArrayList<>(List.of(9L,8L,7L))
         ));
-        HashMap<String, ArrayList<Integer>> m2 = (HashMap<String, ArrayList<Integer>>) deepClone(m1);
+        HashMap<String, ArrayList<Long>> m2 = (HashMap<String, ArrayList<Long>>) deepClone(m1);
         assertEquals(m1, m2);
 
         // Mutate a mutable reference value and ensure no aliasing
-        m1.get("123").add(4);
+        m1.get("123").add(4L);
         assertNotEquals(m1, m2);
     }
 
 
     public static class PTuple_a implements prt.values.PValue<PTuple_a> {
-        public ArrayList<Integer> a;
+        public ArrayList<Long> a;
 
         public PTuple_a() {
-            this.a = new ArrayList<Integer>();
+            this.a = new ArrayList<Long>();
         }
 
-        public PTuple_a(ArrayList<Integer> a) {
+        public PTuple_a(ArrayList<Long> a) {
             this.a = a;
         }
 
         public PTuple_a deepClone() {
-            return new PTuple_a((ArrayList<Integer>)prt.values.Clone.deepClone(a));
+            return new PTuple_a((ArrayList<Long>)prt.values.Clone.deepClone(a));
         } // deepClone()
 
         public boolean equals(Object other) {
@@ -154,12 +154,12 @@ public class ValueCloneTest {
     @Test
     @DisplayName("Can clone a tuple")
     public void testPtupleClone() {
-        PTuple_a t1 = new PTuple_a(new ArrayList<>(List.of(1,2,3)));
+        PTuple_a t1 = new PTuple_a(new ArrayList<>(List.of(1L,2L,3L)));
         PTuple_a t2 = (PTuple_a) deepClone(t1);
 
         assertTrue(t1.deepEquals(t2));
 
-        t1.a.add(99);
+        t1.a.add(99L);
         assertFalse(t1.deepEquals(t2));
         assertNotEquals(t1.a, t2.a);
     }
@@ -167,9 +167,9 @@ public class ValueCloneTest {
     @Test
     @DisplayName("Cannot clone a non-P value class")
     public void testInvalidCloneOfUnrelatedClass() {
-        // AtomicInteger extends j.l.Number extends j.l.Object - this is a totally
+        // AtomicLong extends j.l.Number extends j.l.Object - this is a totally
         // distinct class from any P value the Java code generator emits.
-        AtomicInteger i = new AtomicInteger(42);
+        AtomicLong i = new AtomicLong(42);
         assertThrows(UncloneableValueException.class, () -> deepClone(i));
     }
 
@@ -183,7 +183,7 @@ public class ValueCloneTest {
         // LinkedHashSet extends hashSet, but we expect this should fail nonetheless.
         // (Relaxing this criterion would require walking the inheritance tree via
         // reflection, and it isn't clear what the return type of the cloned value would be.)
-        FancyLinkedHashSet<Integer> lh = new FancyLinkedHashSet<>(List.of(1,2,3,4,5));
+        FancyLinkedHashSet<Long> lh = new FancyLinkedHashSet<>(List.of(1L,2L,3L,4L,5L));
         assertThrows(UncloneableValueException.class, () -> deepClone(lh));
     }
